ver 0.16 - add child element control
powerstrips and dual outlet plugs need to be able to control child
elements. add '-k' option to specify the child. The child can be
specified by ID or by number (which is appended to the parent deviceId).

Also refactored cmd_lookup() to be usable in the module, and added
get_sysinfo() and add_cmd_context() methods.

Removing python2 support